project_root/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
PyMySQL==1.1.1
python-dotenv==1.0.1
bcrypt==4.1.3